Initial Commit. Added all non-mutable files.
Files in ./bin are not changed by runs. Files in ./data that are a part of the
 git project are not changed by runs. Contents of ./data may change during runs
but should be the same at beginning and end of runs.

Not sure if .pyc files should be included in commits, but they're here for now.

Controller.py is the example front-end script, currently set up for a 5-file
example of test scenes from Nina.

BuoyCalib.py is the parent class file, all other classes and files are accessed
through it.

All the other files are named somewhat intelligently, hopefully.