Chnages to narr_data to accomodate new data format (netCDF).
This will allow python to download and handle the data through wget,
without relying on the perl/bash legacy interface. It will be readable, and match the MERRA data acess method well.

This commit includes preliminary updates to download(), open(), and get_points().

Changes will need to be made in choose_points() and read() as well.